+++ MatCalcitePlugin/src/com/github/vlsi/mat/calcite/schema/objects/SpecialFields.java
@@ -1,0 +1,8 @@
+package com.github.vlsi.mat.calcite.schema.objects;
+
+public interface SpecialFields {
+    String CLASS = "@class";
+    String SUPER = "@super";
+    String CLASS_LOADER = "@classLoader";
+    String CLASS_NAME = "@className";
+}